refactor: merge config and sshconfig
